refactor(utils): split `utils` module and remove it
BREAKING CHANGE: `utils` module was removed.
