+++ core/Cargo.toml
@@ -9,36 +9,36 @@ license = "GPL-3.0-only"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "2.0.0-rc", features = [] }
+tauri-build = { version = "2", features = [] }
 
 [dependencies]
-tauri = { version = "2.0.1", features = ["unstable"] }
-tauri-plugin-shell = "2.0.1"
+tauri = { version = "2", features = ["unstable"] }
+tauri-plugin-shell = "2"
 serde = { version = "1", features = ["derive"] }
-serde_json = "1.0"
+serde_json = "1"
 tokio = { version = "1", default-features = false, features = [
   "fs",
   "io-util",
 ] }
 futures = "0.3.21"
-anyhow = "1.0"
-toml = "0.8"
-once_cell = "1.7"
+anyhow = "1"
+toml = "0.9"
+once_cell = "1"
 base64 = "0.22.1"
-regex = { version = "1.11.0", default-features = false, features = ["std"] }
-zip = "2.1.6"
+regex = { version = "1", default-features = false, features = ["std"] }
+zip = "4"
 uuid = { version = "1.4.0", features = ["v4", "fast-rng", "macro-diagnostics"] }
 hematite-nbt = "0.5.2"
-tauri-plugin-http = "2.0.0-rc.0"
-rayon = "1.10.0"
-sha1_smol = "1.0.1"
+tauri-plugin-http = "2"
+rayon = "1"
+sha1_smol = "1"
 log = "0.4.22"
 env_logger = "0.11.5"
-num_cpus = "1.16.0"
-sys-locale = "0.3.1"
+num_cpus = "1"
+sys-locale = "0.3.2"
 tauri-plugin-log = { version = "2", features = ["colored"] }
-tauri-plugin-clipboard-manager = "2.0.2"
-os_info = "3.9.1"
+tauri-plugin-clipboard-manager = "2.3.0"
+os_info = "3.12.0"
 tauri-plugin-dialog = "2"
 backtrace = "0.3.74"
 
@@ -54,4 +54,4 @@ strip = true # Ensures debug symbols are removed.
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
 tauri-plugin-single-instance = "2"
-tauri-plugin-updater = "2.0.0-rc"
+tauri-plugin-updater = "2"
