--- core/src/utils/mod.rs
@@ -1,6 +1,0 @@
-// Conic Launcher
-// Copyright 2022-2026 Broken-Deer and contributors. All rights reserved.
-// SPDX-License-Identifier: GPL-3.0-only
-
-pub mod nbt;
-pub mod unzip;
